Move class to top level, add tests
